remove warnings
[shader]
- X4000: use of potentially uninitialized variable
- X3206: implicit truncation of vector type

[client]
- server warning은 제거하지 않음
- E3251 requires 경고 해결 못함 -> 라이브러리 참조 문제로 추정

+++ Client/X-Engine/shaders/PShader_Deferred.hlsl
@@ -68,7 +68,7 @@ PSOutput_MRT PSDeferred(VSOutput_Standard pin)
     
     if (occlusionMapIndex != -1)
     {
-        occlusion = GammaDecoding(gTextureMaps[occlusionMapIndex].Sample(gsamAnisotropicWrap, pin.UV).x);
+        occlusion = (float)GammaDecoding(gTextureMaps[occlusionMapIndex].Sample(gsamAnisotropicWrap, pin.UV).x);
     }
     
     float rimWidth = 0.8f;
